Feat: train fake news model Refactor: data_cleaning.py -> data_preprocessing.py Del: transformers.py
